+++ src/main/java/com/sprint/mission/discodeit/service/BasicAuthService.java
@@ -1,11 +1,10 @@
-package com.sprint.mission.discodeit.service.basic;
+package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.dto.auth.request.LoginRequest;
-import com.sprint.mission.discodeit.dto.mapper.UserMapper;
-import com.sprint.mission.discodeit.dto.user.data.UserDto;
+import com.sprint.mission.discodeit.dto.auth.LoginRequest;
+import com.sprint.mission.discodeit.mapper.UserMapper;
+import com.sprint.mission.discodeit.dto.user.UserDto;
 import com.sprint.mission.discodeit.entity.main.User;
 import com.sprint.mission.discodeit.repository.UserRepository;
-import com.sprint.mission.discodeit.service.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,12 +12,11 @@ import java.util.NoSuchElementException;
 
 @Service
 @RequiredArgsConstructor
-public class BasicAuthService implements AuthService {
+public class BasicAuthService {
     private final UserRepository userRepository;
 
     private final UserMapper userMapper;
 
-    @Override
     public UserDto login(LoginRequest loginRequest) {
         String username = loginRequest.username();
         String password = loginRequest.password();
