+++ src/main/java/com/sprint/mission/discodeit/service/BasicBinaryContentService.java
@@ -1,10 +1,9 @@
-package com.sprint.mission.discodeit.service.basic;
+package com.sprint.mission.discodeit.service;
 
 import com.sprint.mission.discodeit.dto.binarycontent.BinaryContentDto;
-import com.sprint.mission.discodeit.dto.user.request.BinaryContentCreateRequest;
+import com.sprint.mission.discodeit.dto.user.BinaryContentCreateRequest;
 import com.sprint.mission.discodeit.entity.sub.BinaryContent;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
-import com.sprint.mission.discodeit.service.BinaryContentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,10 +13,9 @@ import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
-public class BasicBinaryContentService implements BinaryContentService {
+public class BasicBinaryContentService {
     private final BinaryContentRepository binaryContentRepository;
 
-    @Override
     public BinaryContentDto create(BinaryContentCreateRequest request) {
         String fileName = request.fileName();
         byte[] bytes = request.bytes();
@@ -31,27 +29,23 @@ public class BasicBinaryContentService implements BinaryContentService {
         return BinaryContentDto.from(binaryContentRepository.save(binaryContent));
     }
 
-    @Override
     public BinaryContentDto find(UUID binaryContentId) {
         return binaryContentRepository.findById(binaryContentId)
                 .map(BinaryContentDto::from)
                 .orElseThrow(() -> new NoSuchElementException("BinaryContent with id " + binaryContentId + " not found"));
     }
 
-    @Override
     public BinaryContent findEntity(UUID id) {
         return binaryContentRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("BinaryContent with id " + id + " not found"));
     }
 
-    @Override
     public List<BinaryContentDto> findAllByIdIn(List<UUID> binaryContentIds) {
         return binaryContentRepository.findAllByIdIn(binaryContentIds).stream()
                 .map(BinaryContentDto::from)
                 .toList();
     }
 
-    @Override
     public void delete(UUID binaryContentId) {
         if (!binaryContentRepository.existsById(binaryContentId)) {
             throw new NoSuchElementException("BinaryContent with id " + binaryContentId + " not found");
