+++ src/main/java/com/sprint/mission/discodeit/service/BasicUserStatusService.java
@@ -1,11 +1,10 @@
-package com.sprint.mission.discodeit.service.basic;
+package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.dto.userstatus.reqeust.UserStatusCreateRequest;
-import com.sprint.mission.discodeit.dto.userstatus.reqeust.UserStatusUpdateRequest;
+import com.sprint.mission.discodeit.dto.userstatus.UserStatusCreateRequest;
+import com.sprint.mission.discodeit.dto.userstatus.UserStatusUpdateRequest;
 import com.sprint.mission.discodeit.entity.sub.UserStatus;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.repository.UserStatusRepository;
-import com.sprint.mission.discodeit.service.UserStatusService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,11 +15,10 @@ import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
-public class BasicUserStatusService implements UserStatusService {
+public class BasicUserStatusService {
     private final UserStatusRepository userStatusRepository;
     private final UserRepository userRepository;
 
-    @Override
     public UserStatus create(UserStatusCreateRequest request) {
         UUID userId = request.userId();
 
@@ -36,19 +34,16 @@ public class BasicUserStatusService implements UserStatusService {
         return userStatusRepository.save(userStatus);
     }
 
-    @Override
     public UserStatus find(UUID userStatusId) {
         return userStatusRepository.findById(userStatusId)
                 .orElseThrow(() -> new NoSuchElementException("UserStatus with id " + userStatusId + " not found"));
     }
 
-    @Override
     public List<UserStatus> findAll() {
         return userStatusRepository.findAll().stream()
                 .toList();
     }
 
-    @Override
     public UserStatus update(UUID userStatusId, UserStatusUpdateRequest request) {
         Instant newLastActiveAt = request.newLastActiveAt();
 
@@ -59,7 +54,6 @@ public class BasicUserStatusService implements UserStatusService {
         return userStatusRepository.save(userStatus);
     }
 
-    @Override
     public UserStatus updateByUserId(UUID userId, UserStatusUpdateRequest request) {
         Instant newLastActiveAt = request.newLastActiveAt();
 
@@ -70,7 +64,6 @@ public class BasicUserStatusService implements UserStatusService {
         return userStatusRepository.save(userStatus);
     }
 
-    @Override
     public void delete(UUID userStatusId) {
         if (!userStatusRepository.existsById(userStatusId)) {
             throw new NoSuchElementException("UserStatus with id " + userStatusId + " not found");
